detector/requirements.txt
pyyaml>=6.0
aiofiles>=23.0
aiohttp>=3.9
psutil>=5.9
flask>=3.0
flask-cors>=4.0
requests>=2.31
